// src/generated/openapi/models/BenchmarkRunDerivedDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV
 * This is an API specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DerivedHostSummaryDTO } from './DerivedHostSummaryDTO';
import {
    DerivedHostSummaryDTOFromJSON,
    DerivedHostSummaryDTOFromJSONTyped,
    DerivedHostSummaryDTOToJSON,
    DerivedHostSummaryDTOToJSONTyped,
} from './DerivedHostSummaryDTO';
import type { DerivedHttpSummaryDTO } from './DerivedHttpSummaryDTO';
import {
    DerivedHttpSummaryDTOFromJSON,
    DerivedHttpSummaryDTOFromJSONTyped,
    DerivedHttpSummaryDTOToJSON,
    DerivedHttpSummaryDTOToJSONTyped,
} from './DerivedHttpSummaryDTO';
import type { DerivedVusSummaryDTO } from './DerivedVusSummaryDTO';
import {
    DerivedVusSummaryDTOFromJSON,
    DerivedVusSummaryDTOFromJSONTyped,
    DerivedVusSummaryDTOToJSON,
    DerivedVusSummaryDTOToJSONTyped,
} from './DerivedVusSummaryDTO';
import type { BenchmarkRunDTO } from './BenchmarkRunDTO';
import {
    BenchmarkRunDTOFromJSON,
    BenchmarkRunDTOFromJSONTyped,
    BenchmarkRunDTOToJSON,
    BenchmarkRunDTOToJSONTyped,
} from './BenchmarkRunDTO';

/**
 * Top-level wrapper for aggregated summary stats for a benchmark run. Percentiles are dynamic via ?percentiles query param (default p90, p95)
 * @export
 * @interface BenchmarkRunDerivedDTO
 */
export interface BenchmarkRunDerivedDTO {
    /**
     * 
     * @type {BenchmarkRunDTO}
     * @memberof BenchmarkRunDerivedDTO
     */
    run?: BenchmarkRunDTO;
    /**
     * HTTP summary grouped per request group
     * @type {Array<DerivedHttpSummaryDTO>}
     * @memberof BenchmarkRunDerivedDTO
     */
    http?: Array<DerivedHttpSummaryDTO>;
    /**
     * 
     * @type {DerivedVusSummaryDTO}
     * @memberof BenchmarkRunDerivedDTO
     */
    vus?: DerivedVusSummaryDTO;
    /**
     * 
     * @type {Array<DerivedHostSummaryDTO>}
     * @memberof BenchmarkRunDerivedDTO
     */
    hosts?: Array<DerivedHostSummaryDTO>;
}

/**
 * Check if a given object implements the BenchmarkRunDerivedDTO interface.
 */
export function instanceOfBenchmarkRunDerivedDTO(value: object): value is BenchmarkRunDerivedDTO {
    return true;
}

export function BenchmarkRunDerivedDTOFromJSON(json: any): BenchmarkRunDerivedDTO {
    return BenchmarkRunDerivedDTOFromJSONTyped(json, false);
}

export function BenchmarkRunDerivedDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BenchmarkRunDerivedDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'run': json['run'] == null ? undefined : BenchmarkRunDTOFromJSON(json['run']),
        'http': json['http'] == null ? undefined : ((json['http'] as Array<any>).map(DerivedHttpSummaryDTOFromJSON)),
        'vus': json['vus'] == null ? undefined : DerivedVusSummaryDTOFromJSON(json['vus']),
        'hosts': json['hosts'] == null ? undefined : ((json['hosts'] as Array<any>).map(DerivedHostSummaryDTOFromJSON)),
    };
}

export function BenchmarkRunDerivedDTOToJSON(json: any): BenchmarkRunDerivedDTO {
    return BenchmarkRunDerivedDTOToJSONTyped(json, false);
}

export function BenchmarkRunDerivedDTOToJSONTyped(value?: BenchmarkRunDerivedDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'run': BenchmarkRunDTOToJSON(value['run']),
        'http': value['http'] == null ? undefined : ((value['http'] as Array<any>).map(DerivedHttpSummaryDTOToJSON)),
        'vus': DerivedVusSummaryDTOToJSON(value['vus']),
        'hosts': value['hosts'] == null ? undefined : ((value['hosts'] as Array<any>).map(DerivedHostSummaryDTOToJSON)),
    };
}