// src/generated/openapi/models/DerivedHttpSummaryDTO.ts
export interface DerivedHttpSummaryDTO {
    /**
     * 
     * @type {string}
     * @memberof DerivedHttpSummaryDTO
     */
    requestGroup?: string;
    /**
     * 
     * @type {string}
     * @memberof DerivedHttpSummaryDTO
     */
    url?: string;
    /**
     * 
     * @type {number}
     * @memberof DerivedHttpSummaryDTO
     */
    totalRequests?: number;
    /**
     * 
     * @type {number}
     * @memberof DerivedHttpSummaryDTO
     */
    requestsPerSecond?: number;
    /**
     * Percentage of requests where httpStatus != expectedStatus
     * @type {number}
     * @memberof DerivedHttpSummaryDTO
     */
    errorRate?: number;
    /**
     * 
     * @type {number}
     * @memberof DerivedHttpSummaryDTO
     */
    totalDataReceivedBytes?: number;
    /**
     * 
     * @type {number}
     * @memberof DerivedHttpSummaryDTO
     */
    totalDataSentBytes?: number;
    /**
     * 
     * @type {DerivedHttpTimingDTO}
     * @memberof DerivedHttpSummaryDTO
     */
    duration?: DerivedHttpTimingDTO;
    /**
     * 
     * @type {DerivedHttpTimingDTO}
     * @memberof DerivedHttpSummaryDTO
     */
    waiting?: DerivedHttpTimingDTO;
}